requirements.txt
transformers
python-dotenv
requests
torch
Image